refactor(settings): split setting-tabs into feature folder
Break down the 1370-line setting-tabs.tsx monolith into a feature-based
structure under src/components/settings/. Each section owns its state via
useSettings() directly; Tauri-related side effects (model download,
autostart, hotkey registration) are extracted into reusable hooks.

New structure:
- settings/SettingTabs.tsx (85 lines) — orchestrator with scroll-spy nav
- settings/common/ — SectionCard, Divider, KeyBadge, SettingsNav, HotkeyInput
- settings/sections/ — Transcription, Audio, Text, Vocabulary, System,
  Shortcuts, Updater (all < 250 lines)
- hooks/useModelDownload, useAutostart, useHotkeyConfig

No behavior change; pnpm build passes.
