+++ docs/REFACTOR_FRONTEND_PROGRESS.md
@@ -10,7 +10,7 @@ Refactoriser le frontend (React + TS) vers une architecture feature-based, avec 
 ## Phases
 
 - [x] **Phase 0** — Setup branche + fichier de suivi + `src/lib/types.ts`
-- [ ] **Phase 1** — Split `setting-tabs.tsx` (1370 → ~300 lignes)
+- [x] **Phase 1** — Split `setting-tabs.tsx` (1370 → 16 fichiers, max 240 lignes, orchestrateur 85 lignes)
 - [ ] **Phase 2** — Split `notes-editor.tsx` (910 → ~250 lignes)
 - [ ] **Phase 3** — Split `Dashboard.tsx` (584 → ~250 lignes) + types centralisés
 - [ ] **Phase 4** — Réorganisation feature folders (dashboard/, notes/, common/, logs/)
@@ -41,6 +41,7 @@ src/
 - `SettingsContext` expose `settings` comme l'objet interne flat (`AppSettings["settings"]`), pas la racine. Les sections peuvent donc utiliser `useSettings()` directement sans props drilling.
 - Chaque phase = 1 commit. Le build (`pnpm build`) doit passer avant commit.
 - Tests UI : l'agent ne peut pas lancer `pnpm tauri dev` — demander à l'utilisateur de valider visuellement après chaque phase.
+- **Phase 1** : les sections consomment `useSettings()` directement ; pas besoin de props drilling depuis `SettingTabs`. Les hooks `useModelDownload`/`useAutostart`/`useHotkeyConfig` encapsulent les invocations Tauri et sont réutilisables hors settings si besoin.
 
 ## Phase 1 — Détail d'extraction (ordre d'exécution)
 
@@ -89,4 +90,25 @@ L'ancien fichier `src/components/setting-tabs.tsx` sera **supprimé** et l'impor
 ## Next session
 
 Prochaine phase : **Phase 2 — notes-editor.tsx split**
-Point de départ : lire `src/components/notes-editor.tsx` complet, créer `src/hooks/useNotesWindow.ts` en premier (logique drag/resize/maximize, lignes ~258-346).
+
+Point de départ :
+1. Lire `docs/REFACTOR_FRONTEND_PROGRESS.md` + dernier commit sur la branche
+2. Lire `src/components/notes-editor.tsx` complet (910 lignes)
+3. Créer `src/hooks/useNotesWindow.ts` en premier (logique drag/resize/maximize/half-screen — lignes ~258-346 de l'ancien fichier ; à re-vérifier après d'éventuels changements)
+4. Puis `useNotesEditor.ts`, `useAiAssistant.ts`, `useLinkEditor.ts`
+5. Puis les composants `NotesEditorWindow/Tabs/Content/Footer` + `EditorBubbleMenu`
+6. Enfin nouveau `NotesEditor.tsx` orchestrateur
+
+Cible : `notes-editor.tsx` 910 → ~200 lignes, aucun nouveau fichier > 250 lignes.
+
+### Phase 1 — Résultats
+
+Fichier supprimé : `src/components/setting-tabs.tsx` (1370 lignes)
+
+Fichiers créés (1505 lignes au total répartis sur 16 fichiers) :
+- `src/components/settings/SettingTabs.tsx` — 85 lignes (orchestrateur)
+- `src/components/settings/common/` — SectionCard (40), Divider (3), KeyBadge (7), SettingsNav (112), HotkeyInput (192)
+- `src/components/settings/sections/` — Transcription (240), Audio (172), Text (99), Vocabulary (149), System (147), Shortcuts (59), Updater (17)
+- `src/hooks/` — useModelDownload (90), useAutostart (38), useHotkeyConfig (55)
+
+Le seul import externe mis à jour : `src/components/Dashboard.tsx` ligne 22.
