+++ src/components/settings/common/Divider.tsx
@@ -1,0 +1,3 @@
+export function Divider() {
+  return <div className="h-px bg-border" />;
+}
