+++ src/components/settings/sections/AudioSection.tsx
@@ -1,0 +1,172 @@
+import { Mic, RefreshCw } from "lucide-react";
+import { Label } from "@/components/ui/label";
+import { Checkbox } from "@/components/ui/checkbox";
+import { Button } from "@/components/ui/button";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { useSettings } from "@/hooks/useSettings";
+import { useAudioDevices } from "@/hooks/useAudioDevices";
+import { SectionCard } from "../common/SectionCard";
+import { Divider } from "../common/Divider";
+
+export function AudioSection() {
+  const { settings, updateSetting } = useSettings();
+  const {
+    devices,
+    isLoading: devicesLoading,
+    error: devicesError,
+    refresh,
+  } = useAudioDevices();
+
+  return (
+    <SectionCard
+      id="section-audio"
+      icon={<Mic className="w-3.5 h-3.5 text-blue-500" />}
+      iconBg="bg-blue-500/10"
+      title="Audio"
+      subtitle="Configuration de l'enregistrement et des sons"
+    >
+      <div className="space-y-5">
+        {/* Microphone */}
+        <div className="space-y-1.5">
+          <div className="flex items-center justify-between">
+            <Label
+              htmlFor="microphone"
+              className="text-sm font-medium text-foreground"
+            >
+              Périphérique d'entrée
+            </Label>
+            <Button
+              variant="ghost"
+              size="sm"
+              onClick={refresh}
+              disabled={devicesLoading}
+              className="h-7 px-2 -mr-1"
+            >
+              <RefreshCw
+                className={`w-3.5 h-3.5 ${devicesLoading ? "animate-spin" : ""}`}
+              />
+            </Button>
+          </div>
+          <Select
+            value={settings.input_device_index?.toString() ?? "null"}
+            onValueChange={(value) =>
+              updateSetting(
+                "input_device_index",
+                value === "null" ? null : Number.parseInt(value),
+              )
+            }
+            disabled={devicesLoading || !!devicesError}
+          >
+            <SelectTrigger id="microphone" className="h-9 bg-background/50">
+              <SelectValue
+                placeholder={
+                  devicesLoading
+                    ? "Chargement..."
+                    : devicesError
+                      ? "Erreur de chargement"
+                      : "Sélectionner un microphone"
+                }
+              />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectItem value="null">Par défaut (Système)</SelectItem>
+              {devices.map((device) => (
+                <SelectItem key={device.index} value={device.index.toString()}>
+                  {device.name} {device.is_default ? "(par défaut)" : ""}
+                </SelectItem>
+              ))}
+            </SelectContent>
+          </Select>
+          {devicesError && (
+            <p className="text-xs text-destructive">Erreur : {devicesError}</p>
+          )}
+        </div>
+
+        {/* Silence threshold */}
+        <div className="space-y-2">
+          <div className="flex items-center justify-between">
+            <Label
+              htmlFor="silence-threshold"
+              className="text-sm font-medium text-foreground"
+            >
+              Seuil de détection du silence
+            </Label>
+            <span className="text-sm font-mono font-semibold text-primary tabular-nums">
+              {(settings.silence_threshold * 100).toFixed(1)}%
+            </span>
+          </div>
+          <input
+            id="silence-threshold"
+            type="range"
+            min="0.001"
+            max="0.05"
+            step="0.001"
+            value={settings.silence_threshold}
+            onChange={(e) =>
+              updateSetting("silence_threshold", parseFloat(e.target.value))
+            }
+            className="w-full h-1.5 bg-muted rounded-full appearance-none cursor-pointer accent-primary"
+          />
+          <p className="text-xs text-muted-foreground">
+            Les enregistrements en dessous de ce seuil seront considérés comme
+            silence
+          </p>
+        </div>
+
+        <Divider />
+
+        {/* Audio options */}
+        <div className="space-y-1">
+          <div
+            className="flex items-center gap-3 p-2.5 rounded-lg hover:bg-muted/30 transition-colors cursor-pointer"
+            onClick={() => updateSetting("enable_sounds", !settings.enable_sounds)}
+          >
+            <Checkbox
+              id="interface-sounds"
+              checked={settings.enable_sounds}
+              onCheckedChange={(checked) =>
+                updateSetting("enable_sounds", checked as boolean)
+              }
+            />
+            <Label
+              htmlFor="interface-sounds"
+              className="text-sm text-foreground cursor-pointer flex-1"
+            >
+              Sons d'interface
+            </Label>
+          </div>
+
+          <div
+            className="flex items-center gap-3 p-2.5 rounded-lg hover:bg-muted/30 transition-colors cursor-pointer"
+            onClick={() =>
+              updateSetting(
+                "enable_history_audio_preview",
+                !settings.enable_history_audio_preview,
+              )
+            }
+          >
+            <Checkbox
+              id="show-listen"
+              checked={settings.enable_history_audio_preview}
+              onCheckedChange={(checked) =>
+                updateSetting("enable_history_audio_preview", checked as boolean)
+              }
+            />
+            <Label
+              htmlFor="show-listen"
+              className="text-sm text-foreground cursor-pointer flex-1"
+            >
+              Bouton Écouter dans l'historique
+            </Label>
+          </div>
+        </div>
+      </div>
+    </SectionCard>
+  );
+}
